Introduce new core module with no junit dependencies

+++ tablasco-core/src/main/java/com/gs/tablasco/verify/indexmap/AdaptivePartialMatcher.java
@@ -41,7 +41,7 @@ public class AdaptivePartialMatcher implements PartialMatcher
         this.actualData = actualData;
         this.expectedData = expectedData;
         this.columnComparators = columnComparators;
-        this.bestMatchThreshold = (long) bestMatchThreshold;
+        this.bestMatchThreshold = bestMatchThreshold;
     }
 
     @Override
